# openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/v1/users/{id}:
    put:
      tags:
      - user-controller
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminUserDTO"
    delete:
      tags:
      - user-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /api/v1/tables/{id}:
    get:
      tags:
      - table-controller
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TableDTO"
    put:
      tags:
      - table-controller
      operationId: update_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TableDTO"
    delete:
      tags:
      - table-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/v1/orders/{id}:
    get:
      tags:
      - order-controller
      operationId: getById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
    put:
      tags:
      - order-controller
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
    delete:
      tags:
      - order-controller
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/v1/menu-items/{id}:
    get:
      tags:
      - menu-item-controller
      operationId: getById_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MenuItemDTO"
    put:
      tags:
      - menu-item-controller
      operationId: update_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MenuItemDTO"
    delete:
      tags:
      - menu-item-controller
      operationId: delete_3
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
  /api/v1/tables:
    get:
      tags:
      - table-controller
      operationId: getAll
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TableDTO"
    post:
      tags:
      - table-controller
      operationId: create
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TableDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TableDTO"
  /api/v1/tables/use/{id}:
    post:
      tags:
      - table-controller
      operationId: use
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
  /api/v1/retellAi/create-appointment:
    post:
      tags:
      - retell-ai-controller
      operationId: createAppointment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentDTO"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/orders:
    get:
      tags:
      - order-controller
      operationId: getAll_1
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderDTO"
    post:
      tags:
      - order-controller
      operationId: create_1
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
  /api/v1/order-items:
    get:
      tags:
      - order-item-controller
      operationId: getAllOrderItems
      parameters:
      - name: orderId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - WAITING
          - PREPARING
          - DONE
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderItemDTO"
    post:
      tags:
      - order-item-controller
      operationId: create_2
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderItemDTO"
  /api/v1/order-items/accept/{id}:
    post:
      tags:
      - order-item-controller
      operationId: accept
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderItemDTO"
  /api/v1/menu-items:
    get:
      tags:
      - menu-item-controller
      operationId: getAllMenuItems
      parameters:
      - name: available
        in: query
        required: false
        schema:
          type: boolean
      - name: category
        in: query
        required: false
        schema:
          type: string
          enum:
          - APPETIZER
          - MAIN_COURSE
          - DESSERT
          - BEVERAGE
          - SIDE
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItemDTO"
    post:
      tags:
      - menu-item-controller
      operationId: create_3
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MenuItemDTO"
  /api/v1/menu-items/{id}/toggle:
    post:
      tags:
      - menu-item-controller
      operationId: toggleAvailability
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MenuItemDTO"
  /api/v1/auth/register:
    post:
      tags:
      - Authentication
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
  /api/v1/auth/password-reset:
    post:
      tags:
      - Authentication
      operationId: resetPassword
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      - name: newPassword
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/auth/forgot-password:
    post:
      tags:
      - Authentication
      operationId: forgotPassword
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/auth/authenticate:
    post:
      tags:
      - Authentication
      operationId: authenticate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthenticationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
  /api/v1/order-items/{id}:
    get:
      tags:
      - order-item-controller
      operationId: getById_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderItemDTO"
    delete:
      tags:
      - order-item-controller
      operationId: delete_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
    patch:
      tags:
      - order-item-controller
      operationId: update_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderItemDTO"
        required: true
      responses:
        "200":
          description: OK
  /api/v1/users/{id}/admin:
    get:
      tags:
      - user-controller
      operationId: getByIdAdmin
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AdminUserDTO"
  /api/v1/users/me:
    get:
      tags:
      - user-controller
      operationId: getMe
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PublicUserDTO"
  /api/v1/users/all:
    get:
      tags:
      - user-controller
      operationId: getAllUsers
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUserDTO"
  /api/v1/tables/waiter:
    get:
      tags:
      - table-controller
      operationId: findForWaiter
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TableDTO"
  /api/v1/tables/order/{id}:
    get:
      tags:
      - table-controller
      operationId: getOrderForTable
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OrderDTO"
  /api/v1/retellAi/date:
    get:
      tags:
      - retell-ai-controller
      operationId: returnDateNow
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
                format: date-time
  /api/v1/orders/recent:
    get:
      tags:
      - order-controller
      operationId: getRecentOrders
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderDTO"
  /api/v1/oauth2/url/google:
    get:
      tags:
      - o-auth-2-controller
      operationId: auth
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/oauth2/authenticate/google:
    get:
      tags:
      - o-auth-2-controller
      operationId: googleAuthenticate
      parameters:
      - name: code
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
  /api/v1/exchange/try:
    get:
      tags:
      - exchange-controller
      operationId: exchangeEuroToTRY
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: number
                format: double
  /api/v1/auth/registrationConfirm:
    get:
      tags:
      - Authentication
      operationId: confirmRegistration
      parameters:
      - name: token
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/v1/auth/refresh-token/{refreshToken}:
    get:
      tags:
      - Authentication
      operationId: refreshToken
      parameters:
      - name: refreshToken
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthenticationResponse"
components:
  schemas:
    AdminUserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - USER
          - ADMIN
          - WAITER
          - COOKER
          - BARMAN
        enabled:
          type: boolean
        password:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          format: date-time
    OrderDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        totalAmount:
          type: number
          format: double
        user:
          $ref: "#/components/schemas/PublicUserDTO"
        orderTime:
          type: string
          format: date-time
        estimatedReadyTime:
          type: string
          format: date-time
        table:
          $ref: "#/components/schemas/TableDTO"
    PublicUserDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - USER
          - ADMIN
          - WAITER
          - COOKER
          - BARMAN
    TableDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: integer
          format: int32
        capacity:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - AVAILABLE
          - OCCUPIED
          - RESERVED
          - NEEDS_CLEANING
        currentOrder:
          $ref: "#/components/schemas/OrderDTO"
        waiter:
          $ref: "#/components/schemas/PublicUserDTO"
    MenuItemDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
          enum:
          - APPETIZER
          - MAIN_COURSE
          - DESSERT
          - BEVERAGE
          - SIDE
        available:
          type: boolean
        preparationTime:
          type: integer
          format: int32
        imageFileName:
          type: string
        image:
          type: string
    AppointmentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        appointmentDate:
          type: string
          format: date-time
        clientName:
          type: string
        clientPhone:
          type: string
    OrderItemDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        menuItem:
          $ref: "#/components/schemas/MenuItemDTO"
        menuItemId:
          type: integer
          format: int64
        name:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
          format: int32
        specialInstructions:
          type: string
        orderItemStatus:
          type: string
          enum:
          - WAITING
          - PREPARING
          - DONE
        order:
          $ref: "#/components/schemas/OrderDTO"
        orderId:
          type: integer
          format: int64
        addedTime:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        surName:
          type: string
        role:
          type: string
          enum:
          - USER
          - ADMIN
          - WAITER
          - COOKER
          - BARMAN
        provider:
          type: string
          enum:
          - LOCAL
          - GOOGLE
    AuthenticationResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: "#/components/schemas/PublicUserDTO"
    AuthenticationRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string